requirements.txt
aiohttp==3.13.4
aiogram==3.26.0
colorama==0.4.6
PyYAML==6.0.3
pydantic==2.12.5